Add GLPITestCase class to be used as parent for all tests
