+++ tests/Autoload.php
@@ -31,5 +31,5 @@
  */
 
 class Autoload extends \DbTestCase {
-   //Placebo class to get atutoloader tested
+   //Placebo class to get autoloader tested
 }
