+++ tests/bootstrap.php
@@ -44,6 +44,7 @@ if (!file_exists(GLPI_CONFIG_DIR . '/config_db.php')) {
 global $CFG_GLPI;
 
 include_once __DIR__ . '/../inc/includes.php';
+include_once __DIR__ . '/GLPITestCase.php';
 include_once __DIR__ . '/DbTestCase.php';
 include_once __DIR__ . '/APIBaseClass.php';
