+++ tests/units/Telemetry.php
@@ -32,8 +32,6 @@
 
 namespace tests\units;
 
-use \atoum;
-
 use \DbTestCase;
 
 /* Test for inc/telemetry.class.php NOT requiring the Web server*/
